refactor(architecture): Decouple SettingsPanel from data logic
Centralizes all data-gathering responsibilities within the ConfigurationService by introducing a new `getWebViewSettings` method.

Previously, the SettingsPanel was tightly coupled to various VS Code APIs and other services to assemble the data payload for the webview.

This commit moves all of that logic into the ConfigurationService, making it the single source of truth for settings data. The SettingsPanel is now a much simpler controller, responsible only for orchestrating communication between the webview and the service.

This improves separation of concerns, enhances testability, and makes the overall architecture cleaner and more maintainable.
